+++ wizard_builder/admin/__init__.py
@@ -1,21 +1,8 @@
 from django.contrib import admin
 
-from ..models import (
-    Checkbox, FormQuestion, MultipleChoice, Page, RadioButton, SingleLineText,
-    TextArea,
-)
+from ..models import FormQuestion, Page
 from .page_admin import PageAdmin
-from .question_admin import (
-    FormQuestionChildAdmin, FormQuestionParentAdmin, MultipleChoiceChildAdmin,
-    MultipleChoiceParentAdmin,
-)
+from .question_admin import FormQuestionAdmin
 
 admin.site.register(Page, PageAdmin)
-
-admin.site.register(FormQuestion, FormQuestionParentAdmin)
-admin.site.register(SingleLineText, FormQuestionChildAdmin)
-admin.site.register(TextArea, FormQuestionChildAdmin)
-
-admin.site.register(MultipleChoice, MultipleChoiceParentAdmin)
-admin.site.register(Checkbox, MultipleChoiceChildAdmin)
-admin.site.register(RadioButton, MultipleChoiceChildAdmin)
+admin.site.register(FormQuestion, FormQuestionAdmin)