+++ wizard_builder/admin/question_admin.py
@@ -1,30 +1,19 @@
 import nested_admin
 
-from .base import ChoiceInline, DowncastedAdmin
+from django import forms
+
+from .. import fields
+from .inlines import ChoiceInline
 
 
-class FormQuestionAdminMixin(object):
-    search_fields = ['text', 'descriptive_text']
+class QuestionAdminForm(forms.ModelForm):
+    type = forms.ChoiceField(choices=fields.get_field_options())
+
+
+class FormQuestionAdmin(
+    nested_admin.NestedModelAdmin,
+):
+    form = QuestionAdminForm
+    search_fields = ['short_str', 'type', 'page']
     list_filter = ['page__sites']
-
-
-class FormQuestionParentAdmin(
-    FormQuestionAdminMixin,
-    DowncastedAdmin
-):
-    list_display = ['short_str', 'model_type', 'page']
-
-
-class FormQuestionChildAdmin(
-    FormQuestionAdminMixin,
-    nested_admin.NestedModelAdmin
-):
-    list_display = ['short_str', 'page']
-
-
-class MultipleChoiceParentAdmin(FormQuestionParentAdmin):
-    inlines = [ChoiceInline]
-
-
-class MultipleChoiceChildAdmin(FormQuestionChildAdmin):
     inlines = [ChoiceInline]