+++ wizard_builder/model_helpers.py
@@ -1,3 +1,13 @@
+class ProxyQuestion:
+
+    def __init__(self, *args, **kwargs):
+        self._meta.get_field('type').default = self.proxy_name
+        super().__init__(*args, **kwargs)
+
+    class Meta:
+        proxy = True
+
+
 class SerializedQuestionMixin:
 
     @property