+++ wizard_builder/tests/test_migrations.py
@@ -74,3 +74,125 @@ class PageIDMigrationTest(MigrationTest):
         new_page_ids = self._get_attrs(NewPage, 'id')
 
         self.assertCountEqual(old_page_ids, new_page_ids)
+
+
+class PopulateTypeMigrationTest(MigrationTest):
+
+    app_name = 'wizard_builder'
+    before = '0028_formquestion_type'
+    after = '0029_populate_type'
+
+    def test_type_populated(self):
+        FormQuestion = self.get_model_before('wizard_builder.FormQuestion')
+        RadioButton = self.get_model_before('wizard_builder.RadioButton')
+        Checkbox = self.get_model_before('wizard_builder.Checkbox')
+        TextArea = self.get_model_before('wizard_builder.TextArea')
+        SingleLineText = self.get_model_before('wizard_builder.SingleLineText')
+
+        formquestion = FormQuestion.objects.create()
+        radiobutton = RadioButton.objects.create()
+        checkbox = Checkbox.objects.create()
+        textarea = TextArea.objects.create()
+        singlelinetext = SingleLineText.objects.create()
+
+        self.run_migration()
+
+        self.assertEqual(
+            FormQuestion.objects.get(id=formquestion.id).type,
+            None,
+        )
+        self.assertEqual(
+            FormQuestion.objects.get(id=radiobutton.id).type,
+            'radiobutton',
+        )
+        self.assertEqual(
+            FormQuestion.objects.get(id=checkbox.id).type,
+            'checkbox',
+        )
+        self.assertEqual(
+            FormQuestion.objects.get(id=textarea.id).type,
+            'textarea',
+        )
+        self.assertEqual(
+            FormQuestion.objects.get(id=singlelinetext.id).type,
+            'singlelinetext',
+        )
+
+
+class PopulateDropdownMigrationTest(MigrationTest):
+
+    app_name = 'wizard_builder'
+    before = '0031_formquestion_choices_default'
+    after = '0032_move_question_dropdown'
+
+    def test_type_populated(self):
+        RadioButton = self.get_model_before('wizard_builder.RadioButton')
+
+        yes_dropdown = RadioButton.objects.create(is_dropdown=True)
+        no_dropdown = RadioButton.objects.create()
+
+        self.run_migration()
+
+        FormQuestion = self.get_model_after('wizard_builder.FormQuestion')
+        yes_question_id = yes_dropdown.formquestion_ptr.id
+        no_question_id = no_dropdown.formquestion_ptr.id
+
+        self.assertEqual(
+            FormQuestion.objects.get(id=yes_question_id).is_dropdown,
+            True,
+        )
+        self.assertEqual(
+            FormQuestion.objects.get(id=no_question_id).is_dropdown,
+            False,
+        )
+
+
+class MoveChoiceQuestionMigrationTest(MigrationTest):
+
+    app_name = 'wizard_builder'
+    before = '0033_add_temps'
+    after = '0034_move_choice_question'
+
+    def test_type_populated(self):
+        RadioButton = self.get_model_before('wizard_builder.RadioButton')
+        OldChoice = self.get_model_before('wizard_builder.Choice')
+
+        question = RadioButton.objects.create()
+        old_choice = OldChoice.objects.create(question=question)
+        old_base_question = old_choice.question.formquestion_ptr
+
+        self.run_migration()
+
+        NewChoice = self.get_model_after('wizard_builder.Choice')
+        new_choice = NewChoice.objects.get(id=old_choice.id)
+        new_base_question = new_choice.question
+
+        self.assertEqual(
+            old_base_question._meta.model_name.lower(), 'formquestion')
+        self.assertEqual(
+            new_base_question._meta.model_name.lower(), 'formquestion')
+        self.assertEqual(
+            old_base_question.id, new_base_question.id)
+
+
+class DropdownMigrationTest(MigrationTest):
+
+    app_name = 'wizard_builder'
+    before = '0039_dropdown_proxy'
+    after = '0040_populate_dropdown'
+
+    def test_type_populated(self):
+        FormQuestion = self.get_model_before('wizard_builder.FormQuestion')
+
+        no_dropdown = FormQuestion.objects.create(
+            type='radiobutton')
+        yes_dropdown = FormQuestion.objects.create(
+            type='radiobutton', is_dropdown=True)
+
+        self.run_migration()
+
+        no_dropdown = FormQuestion.objects.get(id=no_dropdown.id)
+        yes_dropdown = FormQuestion.objects.get(id=yes_dropdown.id)
+
+        self.assertEqual(no_dropdown.type, 'radiobutton')
+        self.assertEqual(yes_dropdown.type, 'dropdown')